+++ pkg/mutator/handler.go
@@ -15,6 +15,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/apimachinery/pkg/types"
+
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 )
 
 type Mutator interface {
@@ -97,7 +99,7 @@ func extractName(request *v1beta1.AdmissionRequest) string {
 	return "<unknown>"
 }
 
-func writeResponse(mutator Mutator, writer http.ResponseWriter, response *v1beta1.AdmissionResponse) {
+func writeResponse(logger generic.Logger, writer http.ResponseWriter, response *v1beta1.AdmissionResponse) {
 	resp, err := json.Marshal(v1beta1.AdmissionReview{
 		TypeMeta: metav1.TypeMeta{
 			Kind:       "AdmissionReview",
@@ -106,11 +108,11 @@ func writeResponse(mutator Mutator, writer http.ResponseWriter, response *v1beta
 		Response: response,
 	})
 	if err != nil {
-		mutator.Log("level", "error", "message", "unable to serialize response", "stack", microerror.JSON(err))
+		logger.Log("level", "error", "message", "unable to serialize response", "stack", microerror.JSON(err))
 		writer.WriteHeader(http.StatusInternalServerError)
 	}
 	if _, err := writer.Write(resp); err != nil {
-		mutator.Log("level", "error", "message", "unable to write response", "stack", microerror.JSON(err))
+		logger.Log("level", "error", "message", "unable to write response", "stack", microerror.JSON(err))
 	}
 }