+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `pkg/filter` package with function which checks if the CR belongs to a cluster from a legacy non-CAPI release.
   Release is considered to be "legacy" if it contains azure-operator.
 - Unit tests for functions from `release` package.
+- `HttpHandlerFactory` for creating HTTP handlers that are using new webhook handlers.
 
 ### Changed